+++ test/helpers/extensions/plugin-api.ts
@@ -14,6 +14,7 @@ export function createTestPluginApi(api: TestPluginApiInput): OpenClawPluginApi 
     registerGatewayMethod() {},
     registerCli() {},
     registerService() {},
+    registerCliBackend() {},
     registerProvider() {},
     registerSpeechProvider() {},
     registerMediaUnderstandingProvider() {},